+++ app/src/main/java/com/academy/android/domain/interactors/GetFilterParametersUseCase.kt
@@ -1,7 +1,7 @@
-package com.academy.android.model.interactors
+package com.academy.android.domain.interactors
 
 import com.academy.android.data.repositories.VideosRepositorySource
-import com.academy.android.model.FilterParameters
+import com.academy.android.domain.models.FilterParameters
 import javax.inject.Inject
 
 class GetFilterParametersUseCase @Inject constructor(
