+++ app/src/main/java/com/academy/android/domain/interactors/NewsLikesInteractor.kt
@@ -1,4 +1,4 @@
-package com.academy.android.model.interactors
+package com.academy.android.domain.interactors
 
 import com.academy.android.data.repositories.NewsRepositorySource
 import javax.inject.Inject
