+++ app/src/main/java/com/academy/android/domain/models/Profile.kt
@@ -1,4 +1,4 @@
-package com.academy.android.model
+package com.academy.android.domain.models
 
 import kotlinx.serialization.Serializable
