+++ app/src/main/java/com/academy/android/ui/news/NewsViewModel.kt
@@ -1,11 +1,11 @@
 package com.academy.android.ui.news
 
 import androidx.lifecycle.ViewModel
-import com.academy.android.model.interactors.GetFeaturedNewsUseCase
-import com.academy.android.model.interactors.GetLikesCountForChatIdUseCase
-import com.academy.android.model.interactors.GetMessagesCountForChatIdUseCase
-import com.academy.android.model.interactors.GetPassedNewsUseCase
-import com.academy.android.model.interactors.NewsLikesInteractor
+import com.academy.android.domain.interactors.GetFeaturedNewsUseCase
+import com.academy.android.domain.interactors.GetLikesCountForChatIdUseCase
+import com.academy.android.domain.interactors.GetMessagesCountForChatIdUseCase
+import com.academy.android.domain.interactors.GetPassedNewsUseCase
+import com.academy.android.domain.interactors.NewsLikesInteractor
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.Flow
